update modal dialogs for bslib
Previously showModal had a scope of a tabpanel.  With bslib, showModal will instantiate each time a navpanel is clicked.  There is only one id for all modals and confusion reigns.  So... now there is only one event for modals with two reactiveVals and lots of if statements to sort it out.